requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Neo4j database
py2neo==2021.2.4

# PDF processing
pdfminer.six==20221105
PyPDF2==3.0.1

# Google AI and embeddings
google-generativeai==0.3.0
python-dotenv==1.0.0

# RAG system and embeddings
langchain==0.0.340
langchain-community==0.0.1
langchain-core==0.1.0
langchain-huggingface==0.0.1
sentence-transformers==2.2.2
faiss-cpu==1.12.0

# Data processing
numpy==1.24.3
pandas==1.5.3

# File uploads and validation
python-multipart==0.0.6

# Additional utilities
requests==2.31.0
aiofiles==23.2.0